Rename package to more generic com.paulmandal.atak.forwarder
