+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.paulmandal.atak.gotenna.mesh">
+    package="com.paulmandal.atak.forwarder">
 
     <application
         android:icon="@mipmap/ic_launcher"
