+++ app/src/main/java/com/paulmandal/atak/forwarder/utils/PermissionUtil.java
@@ -1,4 +1,4 @@
-package com.paulmandal.atak.gotenna.mesh.utils;
+package com.paulmandal.atak.forwarder.utils;
 
 import android.Manifest;
 import android.annotation.TargetApi;
