+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ATAK GoTenna Mesh</string>
+    <string name="app_name">ATAK Forwarder</string>
 </resources>
