+++ app/src/test/java/com/paulmandal/atak/forwarder/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.paulmandal.atak.gotenna.mesh;
+package com.paulmandal.atak.forwarder;
 
 import org.junit.Test;
